Refactor mobile id to xpath func and rename files
